chore: refactor file structure in widgets
