+++ src/app/index.ts
@@ -1,2 +1,1 @@
 export * from "./application";
-export * from "./model";
