+++ src/shared/ui/form-layouts/index.ts
@@ -1,1 +1,1 @@
-export * from "./ui";
+export * from "./view";
