+++ src/widgets/header/constants.ts
@@ -1,6 +1,6 @@
-import { TNavItem } from "shared/ui/nav-item";
+import type { TNavItem } from "~/shared/ui/nav-item";
 
-export const navs: TNavItem[] = [
+export const __NAVS__: TNavItem[] = [
   { id: 1, title: "boards", icon: "common/layers-two", path: "/boards" },
   { id: 2, title: "members", icon: "common/user-circle", path: "/" },
 ];
