+++ src/widgets/header/index.ts
@@ -1,3 +1,1 @@
-export * from "./ui";
-export * as headerModel from "./model";
-export * as headerLib from "./lib";
+export * from "./view";
