--- src/widgets/header/lib/index.ts
@@ -1,2 +1,0 @@
-export * as helpers from "./helpers";
-export * as models from "./models";
