--- src/widgets/header/lib/models.ts
@@ -1,4 +1,0 @@
-export interface IHeaderProps {
-  className?: string;
-  useUser?: boolean;
-}
