+++ src/widgets/header/model.ts
@@ -1,11 +1,11 @@
 import { createEvent, createStore } from "effector";
 
-import { $currentPage } from "shared/routing";
-import { TNavItem } from "shared/ui/nav-item";
+import { $currentPage } from "~/shared/routing";
+import { TNavItem } from "~/shared/ui/nav-item";
 
-import { helpers } from "./lib";
+import { __NAVS__ } from "./constants";
 
-export const $menuItems = createStore<TNavItem[]>(helpers.navs);
+export const $menuItems = createStore<TNavItem[]>(__NAVS__);
 export const $selected = createStore<TNavItem["path"]>("/").on($currentPage, (_, path) => path);
 
 export const navButtonClicked = createEvent<string>();
