+++ src/widgets/page-header/index.ts
@@ -1,2 +1,2 @@
-export * from "./ui";
-export * from "./models";
+export * from "./view";
+export * from "./model";
